refactor(api) : migrate model participant result shared
