+++ api/scripts/prod/compute-participation-results.js
@@ -7,8 +7,8 @@ import * as dotenv from 'dotenv';
 dotenv.config({ path: `${__dirname}/../../.env` });
 import * as campaignRepository from '../../lib/infrastructure/repositories/campaign-repository.js';
 import * as knowlegeElementSnapshotRepository from '../../lib/infrastructure/repositories/knowledge-element-snapshot-repository.js';
+import { ParticipantResultsShared } from '../../src/prescription/campaign-participation/domain/models/ParticipantResultsShared.js';
 import { CampaignParticipationStatuses } from '../../src/prescription/shared/domain/constants.js';
-import { ParticipantResultsShared } from '../../src/shared/domain/models/ParticipantResultsShared.js';
 
 const { SHARED } = CampaignParticipationStatuses;
 import _ from 'lodash';
