+++ api/src/prescription/campaign-participation/domain/models/ParticipantResultsShared.js
@@ -1,6 +1,6 @@
-import { calculatePixScore } from '../../../evaluation/domain/services/scoring/scoring-service.js';
-import { MAX_REACHABLE_PIX_BY_COMPETENCE } from '../constants.js';
-import { KnowledgeElement } from './KnowledgeElement.js';
+import { calculatePixScore } from '../../../../evaluation/domain/services/scoring/scoring-service.js';
+import { MAX_REACHABLE_PIX_BY_COMPETENCE } from '../../../../shared/domain/constants.js';
+import { KnowledgeElement } from '../../../../shared/domain/models/KnowledgeElement.js';
 const MAX_PIX_SCORE = MAX_REACHABLE_PIX_BY_COMPETENCE * 16;
 
 class ParticipantResultsShared {
