+++ api/src/shared/domain/errors.js
@@ -1065,7 +1065,6 @@ export {
   CandidateAlreadyLinkedToUserError,
   CandidateNotAuthorizedToJoinSessionError,
   CandidateNotAuthorizedToResumeCertificationTestError,
-  CantCalculateCampaignParticipationResultError,
   CantImproveCampaignParticipationError,
   CertificateVerificationCodeGenerationTooManyTrials,
   CertificationAttestationGenerationError,
